Corrected slots and signals
Swapped the index and value of the setValue signal and for the recieving slot for XMLParser, DisplayText and Control. Also corrected the functions where they are used.

For the index given to the signal/slot, -1 denotes a value not to be placed on an index, anything greater than or equal to 0 will be placed within the vector of values to be displayed.

+++ include/Display.h
@@ -142,7 +142,7 @@ public slots:
 	void setSpeed(qreal);
 	void setPage(int);
 	[[maybe_unused]] void setValueRange(int);
-    void setValue(double, int);
+    void setValue(int, double);
 private:
     int getDisplayPage(int) const;
 	static QPainterPath addText(QPainterPath, QFont, int, qreal, qreal, const QString&, bool);
